# bo/servlets/govway-apicat-api/src/main/resources/pdnd/openapi_pdnd_client_v3.yaml
...
openapi: 3.0.3
info:
  contact:
    name: API Support
    url: https://github.com/pagopa/pdnd-interop-frontend/issues
  description: 'Sottoinsieme della specifica ufficiale PDND core v3 (m2mGatewayApiV3) contenente le sole
    operazioni invocate da GovCat. L''autenticazione verso la PDND (voucher e DPoP) e'' delegata al gateway
    GovWay: la specifica non dichiara quindi alcuno schema di sicurezza.'
  termsOfService: https://selfcare.interop.pagopa.it/ui/it/termini-di-servizio
  title: PDND Interoperability API (v3) - operazioni utilizzate da GovCat
  version: 3.1.0
  x-summary: PDND Interoperability API
servers:
- url: /v3
  description: PDND core v3
...
paths:
  /status:
    get:
      description: Retrieve the status of the API
      operationId: getStatus
      responses:
        '200':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Status retrieved
          headers:
            Agid-JWT-Signature:
              $ref: '#/components/headers/IntegrityRest02AgidJwtSignatureHeader'
            Digest:
              $ref: '#/components/headers/IntegrityRest02DigestHeader'
            X-Rate-Limit-Interval:
              $ref: '#/components/headers/RateLimitIntervalHeader'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/RateLimitLimitHeader'
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/RateLimitRemainingHeader'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      summary: Retrieve the status of the API
      tags:
      - gateway
  /agreements:
    get:
      description: Retrieve a list of Agreements based on filters
      operationId: getAgreements
      parameters:
      - description: Filter Agreements by states. If not specified, return Agreements in any state
        explode: false
        in: query
        name: states
        required: false
        schema:
          default: []
          items:
            $ref: '#/components/schemas/AgreementState'
          type: array
      - description: Filter Agreements by producer IDs. If not specified, return Agreements for any producer
        explode: false
        in: query
        name: producerIds
        required: false
        schema:
          default: []
          items:
            format: uuid
            type: string
          type: array
      - description: Filter Agreements by consumer IDs. If not specified, return Agreements for any consumer
        explode: false
        in: query
        name: consumerIds
        required: false
        schema:
          default: []
          items:
            format: uuid
            type: string
          type: array
      - description: Filter Agreements by Descriptor IDs. If not specified, return Agreements for any
          Descriptor
        explode: false
        in: query
        name: descriptorIds
        required: false
        schema:
          default: []
          items:
            $ref: '#/components/schemas/DescriptorId'
          type: array
      - $ref: '#/components/parameters/EserviceIdsParam'
      - $ref: '#/components/parameters/OffsetParam'
      - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agreements'
          description: Agreements retrieved
          headers:
            Agid-JWT-Signature:
              $ref: '#/components/headers/IntegrityRest02AgidJwtSignatureHeader'
            Digest:
              $ref: '#/components/headers/IntegrityRest02DigestHeader'
            X-Rate-Limit-Interval:
              $ref: '#/components/headers/RateLimitIntervalHeader'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/RateLimitLimitHeader'
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/RateLimitRemainingHeader'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      summary: List Agreements
      tags:
      - gateway
  /agreements/{agreementId}:
    get:
      description: Retrieve an Agreement
      operationId: getAgreement
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agreement'
          description: Agreement retrieved
          headers:
            Agid-JWT-Signature:
              $ref: '#/components/headers/IntegrityRest02AgidJwtSignatureHeader'
            Digest:
              $ref: '#/components/headers/IntegrityRest02DigestHeader'
            X-Rate-Limit-Interval:
              $ref: '#/components/headers/RateLimitIntervalHeader'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/RateLimitLimitHeader'
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/RateLimitRemainingHeader'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      summary: Retrieve an Agreement
      tags:
      - gateway
      parameters:
      - description: The ID of the Agreement
        in: path
        name: agreementId
        required: true
        schema:
          $ref: '#/components/schemas/AgreementId'
  /agreements/{agreementId}/purposes:
    get:
      description: Retrieve a list of Purposes associated with an Agreement
      operationId: getAgreementPurposes
      parameters:
      - description: The ID of the Agreement to retrieve purposes for
        in: path
        name: agreementId
        required: true
        schema:
          $ref: '#/components/schemas/AgreementId'
      - $ref: '#/components/parameters/LimitParam'
      - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Purposes'
          description: Purposes retrieved for the Agreement
          headers:
            Agid-JWT-Signature:
              $ref: '#/components/headers/IntegrityRest02AgidJwtSignatureHeader'
            Digest:
              $ref: '#/components/headers/IntegrityRest02DigestHeader'
            X-Rate-Limit-Interval:
              $ref: '#/components/headers/RateLimitIntervalHeader'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/RateLimitLimitHeader'
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/RateLimitRemainingHeader'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      summary: List Purposes for an Agreement
      tags:
      - gateway
  /purposes/{purposeId}:
    get:
      description: Retrieve a Purpose
      operationId: getPurpose
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Purpose'
          description: Purpose retrieved
          headers:
            Agid-JWT-Signature:
              $ref: '#/components/headers/IntegrityRest02AgidJwtSignatureHeader'
            Digest:
              $ref: '#/components/headers/IntegrityRest02DigestHeader'
            X-Rate-Limit-Interval:
              $ref: '#/components/headers/RateLimitIntervalHeader'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/RateLimitLimitHeader'
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/RateLimitRemainingHeader'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      summary: Retrieve a Purpose
      tags:
      - gateway
      parameters:
      - description: The ID of the Purpose to retrieve
        in: path
        name: purposeId
        required: true
        schema:
          $ref: '#/components/schemas/PurposeId'
...
  /clients/{clientId}:
    get:
      description: Retrieve a Client by its ID
      operationId: getClient
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
          description: Client retrieved successfully
          headers:
            Agid-JWT-Signature:
              $ref: '#/components/headers/IntegrityRest02AgidJwtSignatureHeader'
            Digest:
              $ref: '#/components/headers/IntegrityRest02DigestHeader'
            X-Rate-Limit-Interval:
              $ref: '#/components/headers/RateLimitIntervalHeader'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/RateLimitLimitHeader'
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/RateLimitRemainingHeader'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      summary: Retrieve a Client
      tags:
      - gateway
      parameters:
      - description: The ID of the Client to retrieve
        in: path
        name: clientId
        required: true
        schema:
          $ref: '#/components/schemas/ClientId'
...
  /eservices/{eserviceId}:
    get:
      description: Retrieve an E-Service
      operationId: getEService
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EService'
          description: E-Service retrieved
          headers:
            Agid-JWT-Signature:
              $ref: '#/components/headers/IntegrityRest02AgidJwtSignatureHeader'
            Digest:
              $ref: '#/components/headers/IntegrityRest02DigestHeader'
            X-Rate-Limit-Interval:
              $ref: '#/components/headers/RateLimitIntervalHeader'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/RateLimitLimitHeader'
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/RateLimitRemainingHeader'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      summary: Retrieve an E-Service
      tags:
      - gateway
      parameters:
      - description: The E-Service ID
        in: path
        name: eserviceId
        required: true
        schema:
          $ref: '#/components/schemas/EServiceId'
  /eservices/{eserviceId}/descriptors:
    get:
      description: Retrieve a list of E-Service Descriptors based on filters
      operationId: getEServiceDescriptors
      parameters:
      - description: The E-Service ID of the Descriptors to retrieve
        in: path
        name: eserviceId
        required: true
        schema:
          $ref: '#/components/schemas/EServiceId'
      - description: Filter Descriptors by states. If not specified, return Descriptors in any state
        in: query
        name: state
        schema:
          $ref: '#/components/schemas/EServiceDescriptorState'
      - $ref: '#/components/parameters/OffsetParam'
      - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EServiceDescriptors'
          description: E-Service Descriptors retrieved
          headers:
            Agid-JWT-Signature:
              $ref: '#/components/headers/IntegrityRest02AgidJwtSignatureHeader'
            Digest:
              $ref: '#/components/headers/IntegrityRest02DigestHeader'
            X-Rate-Limit-Interval:
              $ref: '#/components/headers/RateLimitIntervalHeader'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/RateLimitLimitHeader'
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/RateLimitRemainingHeader'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      summary: List E-Service Descriptors
      tags:
      - gateway
  /eservices/{eserviceId}/descriptors/{descriptorId}:
    get:
      description: Retrieve an E-Service Descriptor
      operationId: getEServiceDescriptor
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EServiceDescriptor'
          description: E-Service Descriptor retrieved
          headers:
            Agid-JWT-Signature:
              $ref: '#/components/headers/IntegrityRest02AgidJwtSignatureHeader'
            Digest:
              $ref: '#/components/headers/IntegrityRest02DigestHeader'
            X-Rate-Limit-Interval:
              $ref: '#/components/headers/RateLimitIntervalHeader'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/RateLimitLimitHeader'
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/RateLimitRemainingHeader'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      summary: Retrieve an E-Service Descriptor
      tags:
      - gateway
      parameters:
      - description: The E-Service ID of the Descriptor
        in: path
        name: eserviceId
        required: true
        schema:
          $ref: '#/components/schemas/EServiceId'
      - description: The ID of the Descriptor
        in: path
        name: descriptorId
        required: true
        schema:
          $ref: '#/components/schemas/DescriptorId'
...
  /certifiedAttributes:
    post:
      description: "Create a certified Attribute\nThe Attribute must\n  - have a unique name across all\
        \ Attributes in the Platform\n  - have a unique code across certified Attributes owned by the\
        \ requester\n"
      operationId: createCertifiedAttribute
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CertifiedAttributeSeed'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertifiedAttribute'
          description: Certified Attribute created
          headers:
            Agid-JWT-Signature:
              $ref: '#/components/headers/IntegrityRest02AgidJwtSignatureHeader'
            Digest:
              $ref: '#/components/headers/IntegrityRest02DigestHeader'
            X-Rate-Limit-Interval:
              $ref: '#/components/headers/RateLimitIntervalHeader'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/RateLimitLimitHeader'
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/RateLimitRemainingHeader'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      summary: Create a certified Attribute
      tags:
      - gateway
  /agreements/{agreementId}/approve:
    post:
      description: Approve an Agreement that is in Pending state, transitioning it to Active state
      operationId: approveAgreement
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agreement'
          description: Agreement approved
          headers:
            Agid-JWT-Signature:
              $ref: '#/components/headers/IntegrityRest02AgidJwtSignatureHeader'
            Digest:
              $ref: '#/components/headers/IntegrityRest02DigestHeader'
            X-Rate-Limit-Interval:
              $ref: '#/components/headers/RateLimitIntervalHeader'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/RateLimitLimitHeader'
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/RateLimitRemainingHeader'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      summary: Approve an Agreement in Pending state
      tags:
      - gateway
      parameters:
      - description: The ID of the Agreement to approve
        in: path
        name: agreementId
        required: true
        schema:
          $ref: '#/components/schemas/AgreementId'
...
components:
  headers:
    IntegrityRest02AgidJwtSignatureHeader:
      description: 'The `Agid-JWT-Signature` header contains a JSON Web Signature (JWS) signed with the
        PDND API private key.

        It ensures payload integrity and non-repudiation across all responses (both success and error),
        in compliance with: 1) Linee Guida sull''interoperabilità tecnica delle Pubbliche Amministrazioni
        - Pattern di sicurezza,  2) RFC 7515 and 3) RFC 7519.

        '
      schema:
        type: string
    IntegrityRest02DigestHeader:
      description: 'Digest of the body using the canonical JSON representation of the response body.

        The digest is calculated using the SHA-256 algorithm on the response body

        exactly as it is sent to the client, encoded in Base64 as per RFC 3230

        '
      schema:
        pattern: ^SHA-256=[a-zA-Z0-9+/]{43}=$
        type: string
    RateLimitIntervalHeader:
      description: Time interval in milliseconds. Allowed requests will be constantly replenished during
        the interval. At the end of the interval the max allowed requests will be available
      schema:
        format: int32
        minimum: 0
        type: integer
    RateLimitLimitHeader:
      description: Max allowed requests within time interval
      schema:
        format: int32
        minimum: 0
        type: integer
...
  schemas:
...
    Problem:
      additionalProperties: false
      properties:
        correlationId:
          description: Unique identifier of the request
          example: 53af4f2d-0c87-41ef-a645-b726a821852b
          maxLength: 64
          type: string
        detail:
          description: A human readable explanation of the problem
          example: Request took too long to complete
          maxLength: 4096
          pattern: ^.{0,1024}$
          type: string
        errors:
          items:
            $ref: '#/components/schemas/ProblemError'
          minItems: 1
          type: array
        status:
          description: The HTTP status code generated by the origin server for this occurrence of the
            problem
          example: 503
          exclusiveMaximum: true
          format: int32
          maximum: 600
          minimum: 100
          type: integer
        title:
          description: A short, summary of the problem type. Written in english and readable
          example: Service Unavailable
          maxLength: 64
          pattern: ^[ -~]{0,64}$
          type: string
        type:
          description: URI reference of type definition
          type: string
      required:
      - type
      - status
      - title
...